Ask the user for the package name and version if needed.
If there is no debian/changelog to find the source package name, and/or
the upstream version was not given on the commandline, ask the user for
them. The values guessed from the original tarball name are offered as
defaults to the user. Previously, this guessed version was used without
confirmation.

The source package name is now always taken from debian/changelog or
prompted for, even when it might not be used (when importing a directory
instead of a tarball, for example). This makes the code more
straightforward and less errorprone in the future.

Closes: #479172