Introduce gbp-pq-rpm
Initial version of gbp-pq-rpm - a tool for managing patch queues for rpm
packages. The functionality more or less corresponds to that of the
(Debian) gbp-pq. The only major difference probably being (in addition
to the obvious of working with .spec files instead of debian/) is that
patches are always imported on top of the upstream version, not on top
of the packaging branch (which might not even contain any source code).

Signed-off-by: Markus Lehtonen <markus.lehtonen@linux.intel.com>
Signed-off-by: Olev Kartau <olev.kartau@intel.com>