PristineTar: move Debian-specific stuff to DebianPristineTar
Continuation to the PristineTar refactoring, makes the "common"
PristineTar independent of DebianPkgPolicy. This commit moves the
Debian-specific has_commit() and checkout() methods to DebianPristineTar
class and replaces them with more generic functions in the base class.
Also, drops the Debian-specific get_commit() method completely, as it
was not used outside the PristineTar class itself.

Signed-off-by: Markus Lehtonen <markus.lehtonen@linux.intel.com>