GitRepository: make get_commit_info() more robust
Now uses git-show instead of git-log. This is needed for further
enhancements (namely to get name-status for merge commits). Also, use
null-character as the field separator which makes parsing more reliable.

The method now returns 'body' of the commit message as is, without
stripping or splitting to lines.

In addition, get_commit_info() now uses GitArgs and _git_inout() instead
of the deprecated _git_getoutput().

Signed-off-by: Markus Lehtonen <markus.lehtonen@linux.intel.com>