rpm-ch: implement tagging options
Implement '--tag' command line option (and other related options for
e.g. signing) for creating and tagging a release. These correspond the
tagging options in gbp-buildpackage-rpm.

A slight difference is that gbp buildpackage-rpm never commits anything
to git. However, in rpm-ch the '--tag' option automatically enables
--commit which causes the changelog modifications (and, all other staged
changes) to be committed to git before creating the tag. This makes it
possible to create a release and document the packaging/release tag name
in the rpm changelog.

Signed-off-by: Markus Lehtonen <markus.lehtonen@linux.intel.com>