Allow git-dch to run in a subdirectory
The --show-cdup flag will output the parent path if "git rev-parse" is being run from a subdirectory. The previous conditional would fail in that case. Git returns a proper 0/128 exit code for "git rev-parse" (tested with versions 1.7.9.5, 1.9.0 and 1.9.1) if it's run outside of a repo, so the `if ret` should be sufficient.