Rewrite gbp.tmpfile
This commit changes the tmpfile module to only contain minimal
init_tmpdir() and del_tmpdir() functions, intended to be called at the
beginning and end of the scipts. The init function creates a base
tempdir and configures the Python tempfile module to use that as the
(default) directory for temporary files and directories.
Correspondingly, the del function removes the base tempdir and restores
the configuration of the tempfile module. The new logic ensures that all
temporary files/directories will be created in the desired base tempdir
(assuming that no 'dir' argument is given to the tempfile functions).

The base tempdir is not removed by del_tmpdir() if a non-empty value of
GBP_TMPFILE_NOCLEAN is found in environment. This can be useful e.g. in
debugging.

Signed-off-by: Markus Lehtonen <markus.lehtonen@linux.intel.com>