Guess upstream-tag based on merge-base
In some unfortunate cases you might have something like this:

$ git tag
v1.0
vyatta/1.0-0something1

... where later tag is closer to describe current HEAD.

With upstream-tag set to v%(version)s find_tag() is going to propose
something based on vyatta/1.0-0something1 which is not expected.

The problem is that find_tag() returns tags, which match the
upstream-tag format, but which are not on the upstream-branch at all.

To fix this find_tag()/git-describe should not be used to determine
the closest upstream-tag.

Closes: gh#7