Allow exported patches to be renumbered and the patch number prefix format to be specified as an option
Now that gbp-pq preserves patch names on import and export the
--[no-]patch-numbers options have been rendered largely useless unless
the patch queue commits are amended with the Gbp[-Pq]: Name tags
removed. The options remain useful if support for renumbering is
added. It's also handy to allow the choice of the format to use
for the numeric prefixes.

Closes: #790890
Signed-off-by: Guido Günther <agx@sigxcpu.org>