Command: redirect stdout/stderr to sys.stdout/stderr
Redirect stdout and stderr of the (child) command to sys.stdout and
sys.stderr of the caller, respectively. This change is mainly for the
unit tests. It makes Python nose to correctly capture the output of the
child command, too, which in turn suppresses a lot of spurious output
when running nosetests.

Closes: #829690

Signed-off-by: Markus Lehtonen <markus.lehtonen@linux.intel.com>