pq: Handle unmerged debian branches (import)
This add two new parameters to pq, upstream_tag, and pq_from, the last one is used to use
the canonical merged branches logic and the use upstream_tag as the base for the pq branch.
This is similar to what's currently used in the rpm code, and could be merged in future
updates.