pq: Parse DEP3 headers
Currently the patch headers in DEP3 format are partially
supported, as git's mailinfo only reads the From and Subject fields from
the first paragraph. But the default in dep3 patches is Description and
Author, that are ignored by git. Even worse, when this fields are in the
first paragraph (again the default) git mailinfo drops all the contained
information.

This patch tries to parse the dep3 headers git's mailinfo couldn't
obtain any useful information, any header other than
Subject|Description and Author|From is appended to the patch message.

The description field is splitted in first line for the short
description and the rest is prepended to the patch message.