import-orig: Move orig.tar.gz with filter-pristine-tar
If we filter a tarball and use filter-pristine-tar we want pristine-tar
to see the correct file name. If a file already exists at that location
move it.

If we wouldn't move the existing tarball but rather e.g. create the
symlink in a temp directory we risk gbp buildpackage picking up the
unfiltered tarball later.

Closes: #558777