git-pbuilder: Check $OPTIONS for --basepath
Previously only the arguments were checked for --basepath.  This
resulted in --basepath being ignored and overridden by the default when
passed in --git-pbuilder-options (either via command arguments or from
the configuration file).  Since this is required when invoking
gbp-buildpackage (because args are passed to --debbuildopts), it's a
significant use case.  This commit fixes the issue by searching both
args and $OPTIONS.

Closes: #856263
Signed-off-by: Kevin Locke <kevin@kevinlocke.name>