New command  'pristine-tar'
This allows one to run the pristine-tar import at a later point.  This
is useful when using component tarballs and we need to reference the
right trees for the commits.

The command can later be extended to allow for "export" as well to
recreate the tarballs without having to invoke "gbp buildpackage".