uscan: Always raise an error if we find <error/> or <warning/>
uscan does not abort with a non-zero exit status when it fails to verify
gpg signatures or similar things so abort when we find a warning or
error.

We can't do any better since uscan stops further processing in this case
of e.g. pgp errors skipping version mangle among other things. See
 #841910.