import-orig: Allow the post-unpack hook to filter tarball files
With post-unpack scripts is currently possible to filter out files that are
not needed from the upstream branch, however it is not possible to use it to
filter files that will end up in the orig file, and in some scenarios this
is not easily doable just using a filter list.

So, run the post-unpack hook just after unpacking and before repacking the
tar.

Add tests to verify this both when using with regular --filter option
and by itself.

Closes #812721