Fix an issue when both export-dir and tarball-dir is set
When the export-dir is set, the git tree is archived and then re-extracted
into the new place.  Unfortunately, different directory is in use to create
the archive (output_dir) and then to extract it (tarball_dir, controlled by
the tarball-dir option).  This means that if no orig.tar.gz is provided in
the tarball_dir directory, the extraction will fail.