repo_setup: Add check/setup_gitattributes
setup_gitattributes() sets up .git/info/attributes in a way to prevent
transformations from interfering with packaging, because the working tree
files can differ from the Git revision history (and from the source packages).

Similar functionality has been implemented by dgit and git-deborig, so we try
to stay compatible and re-use the name of the attribute macro. Since dgit doesn’t
disable export-subst and export-ignore, which may interfere with export-orig, we
add this on top the same way git-deborig does.

dgit used to have an older preset missing some of the attributes, so if
we find those, we disable them.

The user, should they choose so, still can re-enable the attributes
disabled in this way by adding more specific rules further in the
.git/info/attributes file.

Signed-off-by: Andrej Shadura <andrew.shadura@collabora.co.uk>