tests: Implement 'run-in-docker -c docker'
Typically developers have either Docker or Podman on their laptops for
running containers. As Podman follows nearly identical syntax as Docker,
add new option '-c' that allows to specify Podman or Docker as the
container command, with Podman being default if nothing is defined as
the main developer is using Podman.

Also ignore log files both in .gitignore and resulting .dockerignore so
that tests runs can be logged with 'tee x.log' without triggering
rebuilds of the container via polluting the 'COPY . .' step.