docs: Use the shorter URL to online manual, and promote it in man pages too
Shorter is better and prettier, so promote https://gbp.sigxcpu.org/manual/
over the long URL. Also promote this in the introduction of the most
important man pages, as currently the anyone reading the man pages will
not discover the online manual as the links are missing.

The only link in the manpages currently is a local link, which does not
render at all at e.g. https://manpages.debian.org/bookworm/git-buildpackage/git-buildpackage.1.en.html