config: Enable 'multimaint-merge' by default
There are currently 735 packages[1] that default to 'True' for this
option and seems most no maintainer would manually write a
debian/changelog with the same maintainer repeated, so this would be a
better default.

Also, devscripts itself recently changed this and will have it enabled
by default in version 2.25.1+.

This change is also safe to do, as it only affects new debian/changelog
entries and thus there are no backwards compatibility issues.

[1] https://codesearch.debian.net/search?q=multimaint-merge+%3D+True&literal=0
[2] https://salsa.debian.org/debian/devscripts/-/commit/e6b8179080ee087b2a2f3c349b19c6f541335b3b