docs: gbp-clone: Clarify commands and include --add-upstream-vcs in example
Also unify the use of `<screen>` syntax in examples to avoid excess
whitespace, and fix some indentation.

Use `vcs-git:` in examples to avoid confusing reader with multiple options
for the exact same thing. Choose this spelling as it will be correctly
associated with the `Vcs-Git:` field in the `debian/control` file in the
mind of Debian maintainers/developers.

Signed-off-by: Otto Kekäläinen <otto@debian.org>