docs: Add diagram to manual to help readers grasp the DEP-14 branches
To help readers understand what the various branches in DEP-14 typically
are and how they interact, add a diagram to the manual.

This was originally published in my blog at
https://optimizedbyotto.com/post/debian-source-package-git/

I hereby contribute it to git-buildpackage documentation to be used
under the same license I have submitted all my other contributions to
git-buildpackage as well.

Signed-off-by: Otto Kekäläinen <otto@debian.org>