scripts.common.pq: Add some type annotations
datetime.tzinfo is an abstract class with @abstractmethod tzname; mypy
objects to the missing implementation. The return value is
unimportant/arbitrary.

charset.body_encoding is an int; skip the type check there as mypy
doesn't detect it as nullable.

Email headers are a rabbit hole: mypy expects `str` instead of `Header`.
Naive conversion to string results in suboptimal encoding/noise;
correctly converting to a str here requires the msg's policy's `fold`
method. Since this is a legacy (untyped) API, just leave it as is and
ignore the warnings rather than making a subtle mistake.

Signed-off-by: Wesley Hershberger <wesley.hershberger@gmail.com>